refactor(angular): rename addNgPackagr
