+++ packages/angular/src/generators/convert-to-buildable/generators/add-ng-packagr/index.ts
@@ -1,0 +1,1 @@
+export * from './add-ng-packagr';
